Update ruRU.lua
Hi, I have done the translation. But since the volume is quite large, most likely in the future I will make small edits to it (correcting minor inaccuracies, such as the endings of words).

I also added --Дубликат (meaning duplicate) to the text, this is for repeating lines. Is that what you intended? I added them just in case (but they are duplicates).